admin_postman/requirements.txt
aiormq==6.7.7
amqp==5.2.0
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.1
async-timeout==4.0.3
asyncpg==0.29.0
billiard==4.2.0
celery==5.3.6
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
dataclass-wizard==0.22.3
Django==5.0.3
django-celery-results==2.5.1
django-timezone-field==6.1.0
django-tinymce==4.0.0
dnspython==2.6.1
email_validator==2.1.1
exceptiongroup==1.2.0
flower==2.0.1
greenlet==3.0.3
h11==0.14.0
humanize==4.9.0
idna==3.6
Jinja2==3.1.3
kombu==5.3.6
mailhog==0.1.0
MarkupSafe==2.1.5
multidict==6.0.5
orjson==3.10.0
pamqp==3.2.1
prometheus_client==0.20.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pydantic==2.6.4
pydantic-settings==2.2.1
pydantic_core==2.16.3
python-crontab==3.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
schedule==1.2.1
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.4
starlette==0.36.3
tornado==6.4
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.29.0
vine==5.1.0
wcwidth==0.2.13
yarl==1.9.4
gunicorn==21.2.0
